requirements.txt
reflutter==0.7.2
jsbeautifier
pyyaml
colorama
python-magic-bin==0.4.14
lxml
apkid
androguard==3.4.0a1
lz4